+++ arrested/contrib/kim_arrested.py
@@ -23,6 +23,8 @@ class KimHandler(Handler):
             marshaling.
         :param many: Pass the many option to Kim when marshaling or serializing.
         :param raw: Pass the raw option to Kim when marshaling or serializing.
+        :param error_status: Set the HTTP status returned in the event of a
+            MappingInvalid on marshaling
         :returns: None
         """
         super(KimHandler, self).__init__(endpoint, *args, **params)
@@ -34,6 +36,7 @@ class KimHandler(Handler):
         self.raw = params.pop('raw', False)
         self.obj = params.pop('obj', None)
         self.partial = params.pop('partial', False)
+        self.error_status = params.pop('error_status', 422)
 
     @property
     def mapper(self):
@@ -120,10 +123,21 @@ class KimRequestHandler(KimHandler, RequestHandler):
                 )
     """
 
-    def handle(self, data, error_status=422, **kwargs):
-        """Run marshalling for the specified mapper_class.
+    def handle_error(self, exp):
+        """Called if a Mapper returns MappingInvalid. Should handle the error
+        and return it in the appropriate format, can be overridden in order
+        to change the error format.
 
-        * TODO(mike) Allows users to control the error reponse generated.
+        :param exp: MappingInvalid exception raised
+        """
+        payload = {
+            "message": "Invalid or incomplete data provided.",
+            "errors": exp.errors
+        }
+        self.endpoint.return_error(self.error_status, payload=payload)
+
+    def handle(self, data, **kwargs):
+        """Run marshalling for the specified mapper_class.
 
         Supports both .marshal and .many().marshal Kim interfaces.  Handles errors raised
         during marshalling and automatically returns a HTTP error response.
@@ -145,11 +159,7 @@ class KimRequestHandler(KimHandler, RequestHandler):
                     **self.mapper_kwargs
                 ).marshal(role=self.role)
         except MappingInvalid as e:
-            payload = {
-                "message": "Invalid or incomplete data provided.",
-                "errors": e.errors
-            }
-            self.endpoint.return_error(error_status, payload=payload)
+            self.handle_error(e)
 
 
 class KimEndpoint(Endpoint):